REFACTOR: restaurant API implenentations into the restaurantController
